requirements.txt
selenium
fastapi
uvicorn
streamlit
sentence-transformers
faiss-cpu